requirements.txt
flask
gunicorn
supabase
python-dotenv
flask_cors
pandas
plotly
httpx[http2]